# backend/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: backend api
  description: backend api
  version: 1.0.0
paths:
...
  /api/v1/user/{id_user}:
    put:
      tags:
        - user
      parameters:
        - name: id_user
          in: path
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                first_name:
                  type: string
                second_name:
                  type: string
                avatar_url:
                  type: string
                  format: url
                day_of_birth:
                  type: string
                  format: date
                gender:
                  type: string
      responses:
        204:
          description: success update user
        400:
          description: some error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    get:
      tags:
        - user
      parameters:
        - name: id_user
          in: path
          schema:
            type: string
            format: uuid
          required: true
      responses:
        204:
          description: success get user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        400:
          description: some error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /api/v1/user/{id_user}/quiz:
    get:
      tags:
        - quiz
      parameters:
        - name: id_user
          in: path
          schema:
            type: string
            format: uuid
          required: true
        - name: page
          in: query
          schema:
            type: number
        - name: order
          in: query
          schema:
            type: object
            properties:
              date_modify:
                type: string
                enum: [ASC, DESC]
              name:
                type: string
                enum: [ASC, DESC]
              title:
                type: string
                enum: [ASC, DESC]
              description:
                type: string
                enum: [ASC, DESC]
              ttl:
                type: string
                enum: [ASC, DESC]
        - name: description
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
        - name: title
          in: query
          schema:
            type: string
        - name: is_show
          in: query
          schema:
            type: boolean
        - name: is_strict
          in: query
          schema:
            type: boolean
        - name: is_random
          in: query
          schema:
            type: boolean
      responses:
        200:
          description: success get quizzes
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: number
                  quizzes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Quiz'

        400:
          description: some error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      tags:
        - quiz
      parameters:
        - name: id_user
          in: path
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, title]
              properties:
                name:
                  type: string
                title:
                  type: string
                description:
                  type: string
                intro_url:
                  type: string
                ttl:
                  type: string
                tts:
                  type: string
                tte:
                  type: string
                is_show:
                  type: string
                is_strict:
                  type: string
                is_random:
                  type: string
      responses:
        200:
          description: success create quiz
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number

        400:
          description: some error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /api/v1/user/{id_user}/quiz/{id_quiz}:
    get:
      tags:
        - quiz
      parameters:
        - name: id_user
          in: path
          schema:
            type: string
            format: uuid
          required: true
        - name: id_quiz
          in: path
          schema:
            type: number
          required: true
      responses:
        200:
          description: success get quiz
          content:
            application/json:
              schema:
                type: object
                properties:
                  quiz:
                    $ref: '#/components/schemas/Quiz'
        400:
          description: some error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    put:
      tags:
        - quiz
      parameters:
        - name: id_user
          in: path
          schema:
            type: string
            format: uuid
          required: true
        - name: id_quiz
          in: path
          schema:
            type: number
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                title:
                  type: string
                description:
                  type: string
                intro_url:
                  type: string
                ttl:
                  type: string
                tts:
                  type: string
                  format: date
                tte:
                  type: string
                  format: date
                is_show:
                  type: boolean
                is_strict:
                  type: boolean
                is_random:
                  type: boolean
      responses:
        204:
          description: success update quiz
        400:
          description: some error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      tags:
        - quiz
      parameters:
        - name: id_user
          in: path
          schema:
            type: string
            format: uuid
          required: true
        - name: id_quiz
          in: path
          schema:
            type: number
          required: true
      responses:
        204:
          description: success delete quiz
        400:
          description: some error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
  schemas:
...
    Quiz:
      type: object
      properties:
        id:
          type: number
        id_owner:
          type: string
          format: uuid
        name:
          type: string
        title:
          type: string
        description:
          type: string
        intro_url:
          type: string
        date_create:
          type: string
          format: date-time
        date_modify:
          type: string
          format: date-time
        ttl:
          example: 1h10m5s
          description: time to life
          type: string
        tts:
          description: time to start
          type: string
          format: date-time
        tte:
          description: time to end
          type: string
          format: date-time
        is_show:
          type: boolean
        is_strict:
          type: boolean
        is_random:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        first_name:
          type: string
        second_name:
          type: string
        avatar_url:
          type: string
          nullable: true
        day_of_birth:
          type: string
          format: date
          nullable: true
        gender:
          type: string
          nullable: true
    error:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: string
        code:
          type: number
security:
  - cookieAuth: []